requirements.txt
aider-chat==0.67.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
ecdsa==0.19.0
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grep-ast==0.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.53.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
openai==1.56.2
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
playwright==1.47.0
posthog==3.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyee==12.0.0
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.14
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
scipy==1.13.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tiktoken==0.8.0
tokenizers==0.19.1
tqdm==4.67.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
urllib3==2.2.3
watchfiles==1.0.0
wcwidth==0.2.13
Werkzeug==3.1.1
yarl==1.18.3
zipp==3.21.0
